% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CeleriteQFD.R
\name{Celerite}
\alias{Celerite}
\title{Celerite detrending with stan}
\usage{
Celerite(tt, flux, kernel = "Rot", priors = NULL, ...)
}
\arguments{
\item{tt}{the time vector}

\item{flux}{the flux data, make sure there is no NA's}

\item{kernel}{the celerite kernel to be used, currently can only be one of "Rot" for rotation, "SHO" for SHO kernel and "QP" for quasi-periodic kernel. Default "Rot".}

\item{priors}{a list of priors, see details for details on different kernels}

\item{...}{extra argument to stan::sampling}
}
\value{
a list with stan's output in \verb{$stan_out} and trend in \verb{$trend}
}
\description{
Celerite detrending with stan
}
