% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CeleriteQFD-package.R
\docType{package}
\name{CeleriteQFD-package}
\alias{CeleriteQFD-package}
\alias{CeleriteQFD}
\title{The 'CeleriteQFD' package.}
\description{
Simultaneous Detrending and Flare Detection from Photometrical Data using Celerite Gaussian Process with a Quiet-Firing-Decay Hidden Markov Model
}
\references{
Stan Development Team (2020). RStan: the R interface to Stan. R package version 2.21.2. https://mc-stan.org
}
