% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simuFlares.R
\name{kepler_flare}
\alias{kepler_flare}
\title{Simulate n Kepler flares
Routine simulating n Kepler flares in the given time vector tt}
\usage{
kepler_flare(tt, t_half = 5e-05, n = 5, flux_dist = rPareto, ...)
}
\arguments{
\item{tt}{the given time vector where flares were put}

\item{t_half}{the characteristic time scale of the flares}

\item{n}{number of flares to be simulated}

\item{flux_dist}{a function that peak value of the flare was taken, default is Pareto}

\item{...}{extra arguments to flux_dist}
}
\value{
a list with vector of (absolute) flare flux in \verb{$flux} and states (i.e. normal=1, increasing=2 and decay=3) in \verb{$states}
}
\description{
Simulate n Kepler flares
Routine simulating n Kepler flares in the given time vector tt
}
\details{
The characteristic time scale is used for all flares, for each individual flare, its own local scale was the product of its peak value and the overall characteristic timescale. Peak flux was sampled from \code{flux_dist}.
}
