% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simuFlares.R
\name{rPareto}
\alias{rPareto}
\title{Sample from Pareto distribution
Take n samples from the Pareto distribution, kind of power law distribution}
\usage{
rPareto(n, xm = 50, alpha = 1)
}
\arguments{
\item{n}{number of samples}

\item{xm}{lower bound of the distribution, default 50, should be >0}

\item{alpha}{the power it decays, default 1}
}
\value{
a vector of n samples
}
\description{
Sample from Pareto distribution
Take n samples from the Pareto distribution, kind of power law distribution
}
